organiza modulos de firebase para publicaciones
