+++ src/lib/authentication.js
@@ -1,25 +1,21 @@
 import {
-  getAuth,
   createUserWithEmailAndPassword,
   signInWithEmailAndPassword,
   signInWithPopup,
   signOut,
+  GoogleAuthProvider,
 } from 'firebase/auth';
 
-import { auth, provider } from './firebaseConfig';
+import { auth } from './firebaseConfig';
 
 export const signInWithPassword = (email, password) => {
-  // eslint-disable-next-line no-shadow
-  const auth = getAuth();
-  return signInWithEmailAndPassword(auth, email, password);
+  signInWithEmailAndPassword(auth, email, password);
 };
 
 export const registerWithEmail = (email, password) => {
-  // eslint-disable-next-line no-shadow
-  const auth = getAuth();
-  return createUserWithEmailAndPassword(auth, email, password);
+  createUserWithEmailAndPassword(auth, email, password);
 };
 
-export const authGoogle = () => signInWithPopup(auth, provider);
+export const authGoogle = () => signInWithPopup(auth, new GoogleAuthProvider());
 
 export const signOff = () => signOut(auth);
